+++ packages/cli-repl/package.json
@@ -87,7 +87,6 @@
     "js-yaml": "^4.1.0",
     "mongodb-connection-string-url": "^3.0.1",
     "mongodb-log-writer": "^2.3.1",
-    "mongodb-schema": "^12.6.2",
     "numeral": "^2.0.6",
     "pretty-repl": "^4.0.1",
     "semver": "^7.5.4",
